requirements.txt
# fastapi
# uvicorn[standard]
# pyyaml
# numpy
# httpx
# pydantic
# python-dotenv
# openai
# sentence-transformers
# torch


fastapi
uvicorn[standard]
pyyaml
numpy
httpx
pydantic
python-dotenv

# Embeddings stack (pinned to compatible versions)
sentence-transformers==2.2.2
huggingface-hub==0.16.4
torch

